(BSR)[PRO] E2E: add 3 signupJourney tests not yet migrated to cucumber
(BSR)[PRO] E2E: delete useless file

(BSR)[PRO] E2E: more steps after Nicolas review

(BSR)[PRO] E2E: A few typos

(BSR)[PRO] Use initial address, no need to change it

(BSR)[PRO] E2E: intercept changes followinf Amine review
